Map Editing/Loading new maps
Added the V2 Map editor which can save and load maps stored as XML
files. The main game has also added temporarily functionality to load a
default map that must be placed in its persistent data path directory.
Added the default map to the repository. The current map editor does NOT
let the player edit in starting locations, and the starting locations
for players in the main game has been temporarily set to work with only
small sized maps created from the map editor. Not tested with larger
maps.