Map loading
The main game has been updated to read from a directory to find the maps
it needs to load, and creates a directory with the default map if it is
not found. The editor will have this functionality added for saving on
the next version.

Note that the game current loads from ProgramFiles, not Program Files as
I couldn't figure out what unity does for spaces in file path names.