Map Editor v3
The latest version of the map editor should be nearly 100% functional.
It can save and load maps, with an overwrite check for saving. If you
try to load a nonexistant map the program doesnt tell the user anything
but that will be fixed. The only issue I know of so far that can be
easily fixed is the starting locations for player 1 may be read as
player 2 in the game client so I will have to fix that if it is the
case.