GUI additions + Start screen
I left the spells and tech tree GUIs alone. I feel that they simply need
to be moved around a little bit to reorganize the screen. I added
visibility functions to both where needed for the Pause script to work
properly. This also will allow us to turn off other elements on the
screen when accessing the tech tree so as to save space. Pausing is
triggered by pressing 'p' in game.

The start screen scene is roughly implemented. It needs to be made
pretty, but it is functionally working except for loading the map editor
as the editor scene will have to be loaded in and combined before it can
be used.